+++ assembl/static2/css/_assembl_base_styles.scss
@@ -50,7 +50,7 @@
 @import 'components/textWithHeaderPage';
 @import 'components/signup';
 @import 'components/modal';
-@import 'components/embedTheater';
+@import 'components/youtubeEmbed';
 @import 'components/token';
 @import 'components/helper';
 @import 'components/gaugeVoteForProposal';
