--- assembl/static2/css/components/embedTheater.scss
@@ -1,137 +1,0 @@
-.post-section .ReactVirtualized__Grid.ReactVirtualized__List.tree-list {
-  /* Without this, this class uses 'will-change: transform;' which creates
-     a new stacking context and changes the behaviour of 'position: fixed;'
-     childs */
-  will-change: auto !important;
-}
-
-.embed-theater {
-  transition-property: background-color;
-  transition-duration: 2s;
-
-  .embed-video {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-
-    iframe {
-      width: 100%;
-      height: 100%;
-    }
-  }
-
-  .embed-thumbnail-container {
-    width: 100%;
-    position: relative;
-
-    &:hover:after {
-      content: '';
-      display: block;
-      width: 100%;
-      height: 100%;
-      background-color: rgba(0, 0, 0, 0.5);
-      position: absolute;
-      top: 0;
-    }
-
-    .embed-thumbnail {
-      width: 100%;
-      height: auto;
-      position: static;
-    }
-  }
-
-  .play-button {
-    background: transparent;
-    display: inline-block;
-    pointer-events: none;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    position: absolute;
-    z-index: 42;
-
-    &:after {
-      content: '';
-      display: block;
-      width: 0;
-      height: 0;
-      border-style: solid;
-      border-width: 20px 0 20px 40px;
-      border-color: transparent transparent transparent white;
-    }
-  }
-
-  .theater-content {
-    position: relative;
-    width: 100%;
-    height: 0;
-    padding-bottom: 51%;
-    display: block;
-
-    > :first-child {
-      width: 100%;
-      height: 100%;
-      display: block;
-    }
-  }
-
-  &.open {
-    height: 100%;
-    top: 0;
-    left: 0;
-    width: 100%;
-    position: fixed;
-    background-color: rgba(black, 0.8);
-    z-index: 1031;
-
-    .theater-content {
-      width: 90%;
-      top: 50%;
-      left: 50%;
-      transform: translateY(-50%) translateX(-50%);
-
-      > :first-child {
-        height: 100%;
-        width: 100%;
-      }
-
-      .embed-video {
-        max-width: calc(95vh * (16 / 9));
-        max-height: 95vh;
-        transform: translateX(-50%) translateY(-50%);
-        left: 50%;
-        top: 50%;
-      }
-    }
-  }
-
-  .close-theater-button {
-    $size: 30px;
-    $halfSize: $size / 2;
-
-    background-color: black;
-    background-clip: content-box;
-    color: white;
-    position: absolute;
-    border: 0;
-    border: 2px solid white;
-    border-radius: $size / 2;
-    padding: 0;
-    top: -$halfSize;
-    right: -$halfSize;
-    width: $size;
-    height: $size;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    &:before {
-      width: auto;
-      margin: auto;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    }
-  }
-}
