+++ assembl/static2/js/app/components/common/urlPreview/youtubeEmbed.jsx
@@ -1,0 +1,22 @@
+// @flow
+import * as React from 'react';
+
+type Props = {
+  id: string
+};
+
+const YoutubeVideo = ({ id }: Props) => (
+  <div className="video-container">
+    <iframe
+      title="YouTube video"
+      id="ytplayer"
+      type="text/html"
+      width="640"
+      height="360"
+      src={`https://www.youtube.com/embed/${id}?autoplay=1`}
+      frameBorder="0"
+    />
+  </div>
+);
+
+export default YoutubeVideo;
